pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "repl_python3_StaleMeanInterchangeability"
version = "0.1.0"

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
